+++ messages/en-US/cms.php
@@ -5,6 +5,9 @@
  * NOTE: this file must be saved in UTF-8 encoding.
  */
 return [
+    'ASSET_ACTION_DROPDOWN_DELETE' => 'Remove Asset',
+    'ASSET_ACTION_DROPDOWN_DELETE_FILE_MESSAGE' => 'Warning: Deleting this file cannot be undone. All related assets will also be unrecoverably deleted. Please be certain!',
+    'ASSET_ACTION_DROPDOWN_DELETE_MESSAGE' => 'Notice: Removing an asset will not delete the actual file.',
     'ASSET_ALT_TEXT_LABEL' => 'Alt text',
     'ASSET_LINK_LABEL' => 'Link',
     'ASSET_SECTION_ID_LABEL' => 'Section',